refactor: Extract OpsConfigLoader
